Recover DR/DRcloudcage 56.7->96.6%: fn_801EA240 49->98.6% (f32 param not double, f32-LAST callee arg order = left-to-right eval, empty-then clamp, embedded-assign cond, 0.0f literal breaks AE8 CSE), fn_801E9C00 59->96.4% (struct-copy init, walking trail ptr, ctr pair loops w/ unconditional clamp stores, hoisted FP consts in f28-f31, ptr-locals for out-args, points var for r9 reuse, !hidden&&hit guard layout), fn_801EA678 65->92.6% (f32 return, single vtable deref, far-addressed lbl_803AD088)

+++ src/main/dll/DR/DRhightop.c
@@ -619,7 +619,7 @@ undefined4 fn_801EB420(short *param_1,undefined4 param_2,int param_3)
     }
 
     intensity = (int)(lbl_803E6838 * -*(float *)(state + 0x430));
-    fn_801EA240((double)*(float *)(state + 0x49c), (int)param_1, state, intensity,
+    fn_801EA240(*(float *)(state + 0x49c), (int)param_1, state, intensity,
                 state + 0x461, 4);
   }
